backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Celery & Redis
celery==5.3.4
redis==5.0.1

# Storage
minio==7.2.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Payments
stripe==7.4.0

# Logging
loguru==0.7.2

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
